requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.1
typer[all]>=0.12.3
apscheduler>=3.10.4
pytest>=8.2.0
